Extract raft integration into separate file
